+++ example/android/app/src/main/kotlin/com/spencerccf/app_settings_example/MainActivity.kt
@@ -1,7 +1,6 @@
-package com.example.appsettingsexample
+package com.spencerccf.app_settings_example
 
 import io.flutter.embedding.android.FlutterActivity
 
 class MainActivity: FlutterActivity() {
-
 }