+++ ios/Classes/AppSettingsPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface AppSettingsPlugin : NSObject<FlutterPlugin>
-@end